Feature: Exclusion tables - Optionalized params: Now any path params will get populated using the required --onto-name param should they be absent. - CodeStyle: PEP8 compliance update - Script location: Simplified - Ontology config .yml updates: Added missing prefix_map items - Script now utilizes config.yml files for each ontology in order to filter out relevant prefixes. - doid_relevant_signature.sparql: added missing prefix - added warning for excluded terms that appear in relevant_signature file - Now includes non-inclusions, utilizing mirror_signature and component_signature.
