+++ src/ontology/config/omim_exclusion_reasons.robot.template.tsv
@@ -1,0 +1,2 @@
+term_id	exclusion_reason
+ID	AI rdfs:seeAlso
