+++ src/ontology/metadata/doid.yml
@@ -1,4 +1,6 @@
 id: DO
+prefix_map:
+  DOID: http://purl.obolibrary.org/obo/DOID_
 description: >
   The Disease Ontology has been developed as a standardized ontology for 
   human disease with the purpose of providing the biomedical community
