+++ src/ontology/metadata/icd10who.yml
@@ -1,5 +1,7 @@
 id: ICD10WHO
 label: International Statistical Classification of Diseases and Related Health Problems 10th Revision
+prefix_map:
+  ICD10WHO: http://apps.who.int/classifications/icd10/browse/2010/en#/
 description: ICD10WHO is the 10th revision of the International Statistical Classification of Diseases and Related Health Problems (ICD), a medical classification list by the World Health Organization (WHO).
 homepage: https://icd.who.int/browse10/2019/en#/
 base_prefix_map:
