+++ src/ontology/metadata/ncit.yml
@@ -1,5 +1,7 @@
 id: NCIT
 label: NCI Thesaurus (OBO Edition)
+prefix_map:
+  NCIT: http://purl.obolibrary.org/obo/NCIT_
 description: >
   NCI Thesaurus (NCIt)is a reference terminology that includes broad 
   coverage of the cancer domain, including cancer related diseases, findings
