+++ src/ontology/metadata/omim.yml
@@ -1,5 +1,8 @@
 id: OMIM
 label: Online Mendelian Inheritance in Man
+prefix_map:
+   OMIM: https://omim.org/entry/
+   OMIMPS: https://omim.org/phenotypicSeries/PS
 description: >
   OMIM is a comprehensive, authoritative compendium of human genes 
   and genetic phenotypes (https://www.omim.org/about).
