+++ src/ontology/run.sh
@@ -11,7 +11,7 @@
 # See README-editors.md for more details.
 
 IMAGE=${IMAGE:-odkfull}
-ODK_JAVA_OPTS=-Xmx20G
+
 ODK_DEBUG=${ODK_DEBUG:-no}
 
 TIMECMD=
@@ -19,14 +19,7 @@ if [ x$ODK_DEBUG = xyes ]; then
     # If you wish to change the format string, take care of using
     # non-breaking spaces (U+00A0) instead of normal spaces, to
     # prevent the shell from tokenizing the format string.
-    echo "Running ${IMAGE} with ${ODK_JAVA_OPTS} of memory for ROBOT and Java-based pipeline steps."
     TIMECMD="/usr/bin/time -f ### DEBUG STATS ###\nElapsed time: %E\nPeak memory: %M kb"
 fi
 
-docker run -v $PWD/../../:/work -w /work/src/ontology -e ROBOT_JAVA_ARGS="$ODK_JAVA_OPTS" -e JAVA_OPTS="$ODK_JAVA_OPTS" --rm -ti obolibrary/$IMAGE $TIMECMD "$@"
-
-case "$@" in
-*update_repo*|*release*)
-    echo "Please remember to update your ODK image from time to time: https://oboacademy.github.io/obook/howto/odk-update/."
-    ;;
-esac
+docker run -v $PWD/../../:/work -w /work/src/ontology -e ROBOT_JAVA_ARGS='-Xmx20G' -e JAVA_OPTS='-Xmx20G' --rm -ti obolibrary/$IMAGE $TIMECMD "$@"
