+++ src/scripts/update_repo.sh
@@ -28,12 +28,10 @@ mkdir -p $ROOTDIR/.github
 mkdir -p $ROOTDIR/.github/workflows
 cp -n target/$OID/.github/workflows/qc.yml $ROOTDIR/.github/workflows/qc.yml
 
-
-cp target/$OID/.github/workflows/docs.yml $ROOTDIR/.github/workflows/docs.yml
+cp -n target/$OID/.github/workflows/docs.yml $ROOTDIR/.github/workflows/docs.yml
 
 
 cp -n target/$OID/mkdocs.yaml $ROOTDIR/
 
 echo "WARNING: These files should be manually migrated: mkdocs.yaml, .gitignore, src/ontology/catalog.xml (if you added a new import or component)"
-echo "WARNING: Your QC workflows have not been updated automatically. Please update the ODK version number(s) in .github/workflows/qc.yml."
-echo "Ontology repository update successfully completed."
+echo "Update successfully completed."
