+++ src/sparql/get-terms-children.sparql.jinja2
@@ -1,3 +1,6 @@
+{% for sparql_prefix_line in prefixes %}
+{{ sparql_prefix_line }}
+{% endfor %}
 prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#>
 
 select ?term_id ?child_id where {
