+++ src/tmp/robot/component-download-icd10cm-owl-owl__get-terms-children-sparql-jinja2__1/command.sh
@@ -1,0 +1,1 @@
+robot query --input tmp/component-download-icd10cm.owl.owl --query /Users/joeflack4/projects/mondo-ingest/src/scripts/exclusion_term_expansion/cache/robot/component-download-icd10cm-owl-owl__get-terms-children-sparql-jinja2__1/query.sparql /Users/joeflack4/projects/mondo-ingest/src/scripts/exclusion_term_expansion/cache/robot/component-download-icd10cm-owl-owl__get-terms-children-sparql-jinja2__1/results.csv
