+++ src/tmp/robot/icd10cm-ttl__get-terms-children-sparql-jinja2__1/command.sh
@@ -1,0 +1,1 @@
+robot query --input /Users/joeflack4/projects/mondo-ingest/src/scripts/exclusion_term_expansion/../../ontology/mirror/icd10cm.ttl --query /Users/joeflack4/projects/mondo-ingest/src/scripts/exclusion_term_expansion/cache/robot/icd10cm-ttl__get-terms-children-sparql-jinja2__1/query.sparql /Users/joeflack4/projects/mondo-ingest/src/scripts/exclusion_term_expansion/cache/robot/icd10cm-ttl__get-terms-children-sparql-jinja2__1/results.csv
