+++ src/tmp/robot/icd10cm-ttl__get-terms-children-sparql-jinja2__1/query.sparql
@@ -1,0 +1,9 @@
+prefix ICD10CM: <http://purl.bioontology.org/ontology/ICD10CM/>
+prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#>
+prefix skos: <http://www.w3.org/2004/02/skos/core#>
+
+select ?term_id ?child_id ?child_label where {
+  VALUES ?term_id { ICD10CM:P92 ICD10CM:P95 ICD10CM:P96 ICD10CM:P94 ICD10CM:P91 ICD10CM:P93 ICD10CM:P90 }
+  ?child_id rdfs:subClassOf+ ?term_id ;
+    skos:prefLabel ?child_label .
+}
