+++ src/tmp/robot/icd10cm-ttl__get-terms-children-sparql-jinja2__1/results.csv
@@ -1,0 +1,52 @@
+term_id,child_id,child_label
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.9,"Feeding problem of newborn, unspecified"
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.8,Other feeding problems of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.6,Failure to thrive in newborn
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.5,Neonatal difficulty in feeding at breast
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.4,Overfeeding of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.3,Underfeeding of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.2,Slow feeding of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.1,Regurgitation and rumination of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.0,Vomiting of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.09,Other vomiting of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P92,http://purl.bioontology.org/ontology/ICD10CM/P92.01,Bilious vomiting of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.9,"Condition originating in the perinatal period, unspecified"
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.8,Other specified conditions originating in the perinatal period
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.89,Other specified conditions originating in the perinatal period
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.83,Meconium staining
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.82,Delayed separation of umbilical cord
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.81,Exposure to (parental) (environmental) tobacco smoke in the perinatal period
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.5,Complication to newborn due to (fetal) intrauterine procedure
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.3,Wide cranial sutures of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.2,Withdrawal symptoms from therapeutic use of drugs in newborn
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.1,Neonatal withdrawal symptoms from maternal use of drugs of addiction
+http://purl.bioontology.org/ontology/ICD10CM/P96,http://purl.bioontology.org/ontology/ICD10CM/P96.0,Congenital renal failure
+http://purl.bioontology.org/ontology/ICD10CM/P94,http://purl.bioontology.org/ontology/ICD10CM/P94.9,"Disorder of muscle tone of newborn, unspecified"
+http://purl.bioontology.org/ontology/ICD10CM/P94,http://purl.bioontology.org/ontology/ICD10CM/P94.8,Other disorders of muscle tone of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P94,http://purl.bioontology.org/ontology/ICD10CM/P94.2,Congenital hypotonia
+http://purl.bioontology.org/ontology/ICD10CM/P94,http://purl.bioontology.org/ontology/ICD10CM/P94.1,Congenital hypertonia
+http://purl.bioontology.org/ontology/ICD10CM/P94,http://purl.bioontology.org/ontology/ICD10CM/P94.0,Transient neonatal myasthenia gravis
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.9,"Disturbance of cerebral status of newborn, unspecified"
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.8,Other specified disturbances of cerebral status of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.88,Other specified disturbances of cerebral status of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.82,Neonatal cerebral infarction
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.829,"Neonatal cerebral infarction, unspecified side"
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.823,"Neonatal cerebral infarction, bilateral"
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.822,"Neonatal cerebral infarction, left side of brain"
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.821,"Neonatal cerebral infarction, right side of brain"
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.81,Neonatal encephalopathy
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.819,"Neonatal encephalopathy, unspecified"
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.811,Neonatal encephalopathy in diseases classified elsewhere
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.6,Hypoxic ischemic encephalopathy [HIE]
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.63,Severe hypoxic ischemic encephalopathy [HIE]
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.62,Moderate hypoxic ischemic encephalopathy [HIE]
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.61,Mild hypoxic ischemic encephalopathy [HIE]
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.60,"Hypoxic ischemic encephalopathy [HIE], unspecified"
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.5,Neonatal coma
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.4,Neonatal cerebral depression
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.3,Neonatal cerebral irritability
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.2,Neonatal cerebral leukomalacia
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.1,Acquired periventricular cysts of newborn
+http://purl.bioontology.org/ontology/ICD10CM/P91,http://purl.bioontology.org/ontology/ICD10CM/P91.0,Neonatal cerebral ischemia
+http://purl.bioontology.org/ontology/ICD10CM/P93,http://purl.bioontology.org/ontology/ICD10CM/P93.8,Other reactions and intoxications due to drugs administered to newborn
+http://purl.bioontology.org/ontology/ICD10CM/P93,http://purl.bioontology.org/ontology/ICD10CM/P93.0,Grey baby syndrome
